day-64/requirements.txt
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
WTForms==3.0.1
requests==2.28.1